Remove LoginEpilogue
